% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainlinkDF.R
\name{chainlinkDF}
\alias{chainlinkDF}
\title{Get a data frame of chained volumes}
\usage{
chainlinkDF(df_list, ref_year = 2015)
}
\arguments{
\item{df_list}{A list, where the list elements are each a data frame consisting of current price (cup) and previous year price (pyp) values.}

\item{ref_year}{Reference year for Annual Overlap. Defaults to 2015.}
}
\value{
a data.frame of chained volumes.
}
\description{
Get a data frame of chained volumes
}
