% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contGrowth.R
\name{contGrowth}
\alias{contGrowth}
\title{Contributions to Growth}
\usage{
contGrowth(x, prices = "nom", baseline_var = NULL, ref_lag = NULL)
}
\arguments{
\item{x}{A list of equally structured Time Series Data Frames. For nom contribution to growth the list is not nessecary it can also be a time series data frame.}

\item{prices}{selects nominal growth rates ("nom") or chain-linked volumes ("vol"), Defaults to "nom".}

\item{baseline_var}{Name of the variable of which growth rates should add up to. Usually "BIP". If NULL it looks for variables that are called BIP or GDP, otherwise it takes the first variable.}

\item{ref_lag}{lag to reference period. defaults to frequency of the series, i.e. yoy growth rates. To get qoq growth rates change to 1.}
}
\value{
A Time series object w/ the same dimensions as the input list elements
}
\description{
Compile contribution to year-on-year growth rates from a list of Time Series.
}
\details{
As chain-linked volumes are not additive, values have to be adapted to be additive.
The Additive Volume Data (AVD) Method, where weights are derived from deflators is recommended by EUROSTAT
\deqn{C.AVD(x_t, X)^{c, y-1 \rightarrow c, y} = \frac{P_{x_t}^{y-1}}{P_{X}^{y-1}}\frac{(x_{CVt}^{c,y}- x_{CVt}^{c,y-1})}{X_{CV}^{c,y-1}} +   (\frac{x_{CVt}^{c,y-1}}{X_{CV}^{c,y-1}}-\frac{x_{CVt}^{y-1}}{X_{CV}^{y-1}})*(\frac{P_{x_t}^{y-1}}{P_{X}^{y-1}}-\frac{P_{x_t}^{y-2}}{P_{X}^{y-2}})}
}
