% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthRate.R
\name{growthRate}
\alias{growthRate}
\title{Growth Rates}
\usage{
growthRate(x, mode = "additive", ref_lag = NULL)
}
\arguments{
\item{x}{A time series (uni- or multivariate).}

\item{mode}{additive or multiplicative growth rates. Defaults to additive.}

\item{ref_lag}{lag to reference period. defaults to frequency of the series, i.e. yoy growth rates. To get qoq growth rates change to 1.}
}
\value{
time series with growth rates
}
\description{
Growth Rates
}
