% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsToDf.R
\name{tsToDf}
\alias{tsToDf}
\title{Convert time series to dataframe}
\usage{
tsToDf(x, lang = "en")
}
\arguments{
\item{x}{A time series dataframe without colums indicating the time period}

\item{lang}{Language. Possible values: "de" for German or  "en" for englisch. Defaults to "en"}
}
\value{
a data. frame with (one or) two additional columns to indicate YEAR and QUARTER/MONTH if frequency of input tiome series is 4 or 12 respectively.
}
\description{
Convert time series to dataframe
}
