% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainlinkAO.R
\name{chainlinkAO}
\alias{chainlinkAO}
\title{Creates chainlinked time series with annual overlap}
\usage{
chainlinkAO(cup, pyp, ref_year = 2015, index = F)
}
\arguments{
\item{cup}{univariate time series at current prices (cup)}

\item{pyp}{univariate time series at previous year prices (pyp)}

\item{ref_year}{Base year. Defaults to 2010}

\item{index}{If TRUE it returns an index series, where the base year is 100, if FALSE returns volume series. Defaults to FALSE.}
}
\value{
a univariate time series with volume or index values
}
\description{
Using the annual overlap approach this function chainlinks univariate time series with respect to the given base year. Frequency of time series can also be greater than 1.
}
