% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chainlinkDF.R
\name{chainlinkDF}
\alias{chainlinkDF}
\title{Erstellt einen Data Frame mit veketten Volumen}
\usage{
chainlinkDF(df_list, ts_output = F, basis = 2015)
}
\arguments{
\item{df_list}{Eine Liste, bei der die Elemente jeweils ein Data Frame sind mit nominellen Werten bzw. Werten zu VJP.}

\item{ts_output}{logical. Soll der Output eine Zeitreihe sein, oder nicht. Defaults to FALSE}

\item{basis}{Referenzjahr für Verkettung. Default ist 2015}
}
\value{
einen data.table, die Werte sind verkette Volumen
}
\description{
Erstellt einen Data Frame mit veketten Volumen
}
