% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsToDf.R
\name{tsToDf}
\alias{tsToDf}
\title{Convert TS to dataframe}
\usage{
tsToDf(x, lang = "en")
}
\arguments{
\item{x}{A time series dataframe without colums indicating the time period}

\item{lang}{Language german ("de") or englisch "en" colnames (YEAR, QUARTER). Defaukts to "en"}
}
\value{
a data. frame
}
\description{
Convert TS to dataframe
}
